disable box drops - DO NOT USE THIS
THIS WILL LEAK MEMORY, as we are not properly dropping boxes anymore.

Time and time again, we have found problems with the box-driven ref-tracking
and the proper dropping of data. This often shows up in Flutter assuming
objects to be alive, whose memory has even been overwritten already and thus
related functions turn into garbage calls - most notably in String decoding.

In this latest installment we've found that Task were dropped on the rust-side
when the outer FfiListTask was gone, even though we still needed to read the
title and other data. This shows up a lot more clearly on flutter profile
or release mode, leading into a strong crash.

In this patch, which should never go into mainline, we are hard-core disabling
the box drop calls to the rust side, effectively leaking all memory that we've
ever received from rust.