Add 32bit support, make generated dart portable
This replaces several instance where we were using u64 or i64 as hard-coded
instances for isize and usize with its dart counterparts of `UintPtr` and `IntPtr``.
Thus rendering the generated dart file not only platform-compatible with
the underlying target system, but also allowing us to the same generated
dart file across any number of platforms. Prior to this for the few cases
this was using usize-types, the generator would create a platform-specific
mapping.

As a result, we do now support 32bit targets as well.